Make eligibility checks more async
